% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunction.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion}
\usage{
bic(ttnsr,theta,omega,d,r)
}
\arguments{
\item{ttnsr}{Observed tensor data}

\item{theta}{continuous parameter tensor (latent parameter)}

\item{omega}{the cut-off points (k-levels)}

\item{d}{Dimension of tensor}

\item{r}{Rank of tensor}
}
\value{
BIC value given observed tensor, estimated parameter and rank
}
\description{
Bayesian Information Criterion evaluated at given parameters, dimension of tensor and rank of tensor
}
