% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunction.R
\name{estimation}
\alias{estimation}
\title{Predict the entries of a tensor from the cumulative model}
\usage{
estimation(theta,omega,type = c("mode","mean","median"))
}
\arguments{
\item{theta}{continuous parameter tensor (latent parameter)}

\item{omega}{the cut-off points (k-levels)}

\item{type}{type of predictions:

\code{"mode"} specifies argmax based label prediction

\code{"mean"} specifies mean based label prediction

\code{"median"} specifies median based label prediction}
}
\value{
predicted ordinal tensor from the given parameters and types of prediction
}
\description{
Predict the ordinal-valued entries of a tensor with given parameters and a type of predictions
}
\examples{
indices <- c(10,20,30)
arr <- array(runif(prod(indices)),dim = indices)
b <- runif(3)
r_predict <- estimation(arr,b,type = "mode");r_predict
}
