% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunctions.R
\name{fit_continuous}
\alias{fit_continuous}
\title{Fit a tensor using Tucker model}
\usage{
fit_continuous(ttnsr,r,alpha = TRUE)
}
\arguments{
\item{ttnsr}{an observed tensor data}

\item{r}{a rank to be fitted (Tucker rank)}

\item{alpha}{a max-norm constraint of a tensor

\code{alpha = TRUE} if there is no constraint for max-norm of a latent tensor parameter}
}
\value{
a list containing the following:

\code{C} - an estimated core tensor

\code{A} - estimated factor matrices

\code{iteration} - the number of iterations

\code{cost} - evaluations of cost functions at each iterations
}
\description{
Fitting e a tensor with possibly missing values using Tucker model
}
\examples{
# Latent parameters
library(rTensor)
alpha = 10
A_1 = matrix(runif(15*2,min=-1,max=1),nrow = 15)
A_2 = matrix(runif(15*2,min=-1,max=1),nrow = 15)
A_3 = matrix(runif(15*2,min=-1,max=1),nrow = 15)
C = as.tensor(array(runif(2^3,min=-1,max=1),dim = c(2,2,2)))
theta = ttm(ttm(ttm(C,A_1,1),A_2,2),A_3,3)@data
theta = alpha*theta/max(abs(theta))
omega = c(-0.2,0.2)

# Observed tensor
ttnsr <- realization(theta,omega)@data

# Estimation of parameters
continuous_est = fit_continuous(ttnsr,c(2,2,2),alpha = 100)

}
