% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunction.R
\name{likelihood}
\alias{likelihood}
\title{Log-likelihood function (cost function)}
\usage{
likelihood(ttnsr,theta,omega,type = c("ordinal","Gaussian"))
}
\arguments{
\item{ttnsr}{observed tensor}

\item{theta}{continuous parameter tensor (latent parameter)}

\item{omega}{the cut-off points (k-levels)}

\item{type}{types of log-likelihood function

\code{"ordinal"} specifies log-likelihood function with the cumulative logistic model

\code{"Gaussian"} specifies squared error function}
}
\value{
a value evaluated on given type and points
}
\description{
Log-likelihood function (cost function)
}
