% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunction.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC) value}
\usage{
bic(ttnsr,theta,omega,d,r)
}
\arguments{
\item{ttnsr}{an observed tensor data}

\item{theta}{continuous parameter tensor (latent parameter)}

\item{omega}{the cut-off points}

\item{d}{dimension of the tensor}

\item{r}{rank of the tensor}
}
\value{
BIC value at given inputs
}
\description{
Obtain Bayesian Information Criterion (BIC) evaluated at a given parameter tensor, an observed tensor, dimension and rank of the tensors
}
