% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunction.R
\name{estimation}
\alias{estimation}
\title{Estimate tensor entries from the cumulative model}
\usage{
estimation(theta,omega,type = c("mode","mean","median"))
}
\arguments{
\item{theta}{continuous parameter tensor (latent parameter)}

\item{omega}{the cut-off points}

\item{type}{type of estimations:

\code{"mode"} specifies argmax based label estimation

\code{"mean"} specifies mean based label estimation

\code{"median"} specifies median based label estimation}
}
\value{
an estimated ordinal tensor from latent parameters and types of estimations
}
\description{
Estimating the ordinal-valued tensor entries from latent parameters and a type of estimations
}
\examples{
indices <- c(10,20,30)
arr <- array(runif(prod(indices)),dim = indices)
b <- qnorm((1:3)/4)
r_predict <- estimation(arr,b,type = "mode");r_predict
}
