% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunctions.R
\name{fit_continuous}
\alias{fit_continuous}
\title{Continuous Tucker decomposition}
\usage{
fit_continuous(ttnsr,C,A_1,A_2,A_3,alpha = TRUE)
}
\arguments{
\item{ttnsr}{tensor data (k-level)}

\item{C}{initial point for a core tensor and factor matrices}

\item{A_1, A_2, A_3}{initial points for factor matrices}

\item{alpha}{max-norm constraints of a tensor

\code{alpha = TRUE} if there is no constraint for max-norm of latent tensor parameters}
}
\value{
a list containing the following:

\code{C} - estimated core tensor

\code{A_1, A_2, A_3} - estimated factor matrices

\code{iteration} - the number of iterations

\code{cost} - evaluations of cost functions at each iterations
}
\description{
decompose a tensor with possibly missing values into Tucker decomposition
}
\examples{
# draw random ordinal tensor
indices = c(10,10,10)
rank = c(3,3,3)
otensor <- array(sample(c(1,2,3),prod(indices),replace = TRUE),dim = indices)

#Initial points for `fit_continuous' function
A_1 = pracma::randortho(indices[1])[,1:rank[1]]
A_2 = pracma::randortho(indices[2])[,1:rank[2]]
A_3 = pracma::randortho(indices[3])[,1:rank[3]]
C = rTensor::rand_tensor(modes = rank)

#Tucker decomposition
Tucker_est = fit_continuous(otensor,C,A_1,A_2,A_3,alpha = 100)
}
