% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunctions.R
\name{fit_ordinal}
\alias{fit_ordinal}
\title{Fit the cumulative logistic model}
\usage{
fit_ordinal(ttnsr,r,omega=TRUE,alpha = TRUE)
}
\arguments{
\item{ttnsr}{an observed tensor data}

\item{r}{a rank to be fitted (Tucker rank)}

\item{omega}{cut-off points if it is known,

\code{omega = TRUE} if it is unknown}

\item{alpha}{a signal level

\code{alpha = TRUE} if there is no constraint for max-norm of latent tensor parameters}
}
\value{
a list containing the following:

\code{C} - an estimated core tensor

\code{A} - estimated factor matrices

\code{theta} - an estimated latent parameter tensor

\code{iteration} - the number of iterations

\code{cost} - evaluations of cost functions at each iterations

\code{omega} - estimated cut-off points vector
}
\description{
Fitting the cumulative logistic model with alternating optimization method.
}
\examples{
# Latent parameters
library(rTensor)
alpha = 10
A_1 = matrix(runif(15*2,min=-1,max=1),nrow = 15)
A_2 = matrix(runif(15*2,min=-1,max=1),nrow = 15)
A_3 = matrix(runif(15*2,min=-1,max=1),nrow = 15)
C = as.tensor(array(runif(2^3,min=-1,max=1),dim = c(2,2,2)))
theta = ttm(ttm(ttm(C,A_1,1),A_2,2),A_3,3)@data
theta = alpha*theta/max(abs(theta))
omega = c(-0.2,0.2)

# Observed tensor
ttnsr <- realization(theta,omega)@data

# Estimation of parameters
ordinal_est = fit_ordinal(ttnsr,c(2,2,2),omega = TRUE,alpha = 100)

}
\references{
Lee and Wang (2020) <arXiv:2002.06524>.
}
