% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfunction.R
\name{likelihood}
\alias{likelihood}
\title{Obtain log-likelihood function (cost function) value}
\usage{
likelihood(ttnsr,theta,omega,type = c("ordinal","Gaussian"))
}
\arguments{
\item{ttnsr}{an observed tensor data}

\item{theta}{continuous parameter tensor (latent parameter)}

\item{omega}{the cut-off points}

\item{type}{types of log-likelihood function

\code{"ordinal"} specifies log-likelihood function with the cumulative logistic model

\code{"Gaussian"} specifies log-likelihood function with the Gaussian model}
}
\value{
log-likelihood value at given inputs
}
\description{
Log-likelihood function (cost function)
}
